requirements.txt
numpy
scipy
matplotlib
voila
pandas
sympy
nose
python-telegram-bot